moved setupEnvVariables to it's own file
